% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addIPgeolocation}
\alias{addIPgeolocation}
\title{------------------------------------GEOLOCALIZACIÓN DE DATOS
Funcion para obtención de geolocalizacion de las IPs}
\usage{
addIPgeolocation(ips = df.muestra$saddr, df.geoip, boost = FALSE)
}
\arguments{
\item{ips}{arreglo de caracteres correspondientes a direcciones IPv4}

\item{df.geoip}{dataset descargado con la función downloadScanIO}

\item{boost}{variable por defecto inicializada como FALSE. Si es TRUE usará
parallel computing using multiple cores.}
}
\value{
dataset df
}
\description{
------------------------------------GEOLOCALIZACIÓN DE DATOS
Funcion para obtención de geolocalizacion de las IPs
}
\examples{
\dontrun{
 geo.ips <- addIPgeolocation(ips = df.muestra$saddr, df.geoip = download.geoip())
 geo.ipd <- addIPgeolocation(ips = df.muestra$daddr, df.geoip = download.geoip())

}


}
