% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Adiciona.pais}
\alias{Adiciona.pais}
\title{Función para adicionar el nombre de país que nos servirá para el posterior análisis
Se adiciona la columna PAIS al dataset}
\usage{
Adiciona.pais(geo)
}
\arguments{
\item{geo}{En este parámetro se debe ingresar el resultado de la función addIPgeolocation
A partir de ella se puede trabajar con IPs de origen o destino}
}
\value{
data set más columna de paises obtenida a partir de longitud, latitud
}
\description{
Función para adicionar el nombre de país que nos servirá para el posterior análisis
Se adiciona la columna PAIS al dataset
}
\examples{
 \dontrun{
geo.ips <- Adiciona.pais(geo.ips)
geo.ipd <- Adiciona.pais(geo.ipd)

}


}
