% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{CrearDirectorio}
\alias{CrearDirectorio}
\title{-------------------------------------CREAR FOLDER DE TRABAJO
Función para crear directorio de trabajo}
\usage{
CrearDirectorio(n.folder = "datasets")
}
\arguments{
\item{n.folder}{Define nombre de folder o carpeta donde se almacenarán datasets}

\item{verbose}{Variable de inicialización}
}
\value{
Devuelve directorio donde se almacenarán datasets
}
\description{
Package para creación de directorio.
}
\details{
(tini) variable donde almacenams la hora del sistema
(n.folder) Directorio donde iremos alnacenar los ficheros descargados.
}
\author{
Cristiano Dias / Luiggi Alexis Rodriguez Ruiz
}
