% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{download.geoip}
\alias{download.geoip}
\title{------------------------------------------DOWNLOAD 2
Funcion para descargar y descomprimir el dataset MAxmind para obtener las longitudes y
latitudes.
La función incluye la validación de la creación de la carpeta destino}
\usage{
download.geoip(n.folder = "datasets")
}
\arguments{
\item{n.folder}{Nombre del folder donde se almacenará dataset}
}
\value{
Devuelve dataset descargado con los datos en crudo
}
\description{
------------------------------------------DOWNLOAD 2
Funcion para descargar y descomprimir el dataset MAxmind para obtener las longitudes y
latitudes.
La función incluye la validación de la creación de la carpeta destino
}
