% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{downloadScanIO}
\alias{downloadScanIO}
\title{Funcion para descargar y descomprimir}
\usage{
downloadScanIO(data.url, dir.path = "dados3", filename)
}
\arguments{
\item{data.url}{Introducir la url con el dataset en formato csv que queremos descargar.}

\item{dir.path}{Nombre del directorio}

\item{filename}{Fichero que tendrá el dataset}
}
\value{
Devuelve el dataframe descargado con los datos en crudo
}
\description{
Funcion para descargar y descomprimir
}
\examples{
url="https://opendata.rapid7.com/sonar.tcp/2019-04-20-1555725774-https_get_16993.csv.gz"

downloadScanIO(url,"http","scansio.http")

dataset=downloadScanIO(url,"http","scansio.http")  # Almacena en la variable dataset

\dontrun {
downloadScanIO(url,"http","scansio.http")
}
\dontrun {
dataset=downloadScanIO(url,"http","scansio.http")
}
}
