% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{filename}
\alias{filename}
\title{Title Función para crear directorio de trabajo}
\format{An object of class \code{character} of length 1.}
\usage{
filename
}
\arguments{
\item{verbose}{valor TRUE}
}
\description{
Package para creación de directorio.
}
\details{
(tini) variable donde almacenams la hora del sistema
(dir.path) Directorio donde iremos alnacenar los ficheros descargados.
}
\examples{
\dontrun{
CrearDirectorio("prueba")
CrearDirectorio()
}
}
\author{
Cristiano Dias / Luiggi Alexis Rodriguez Ruiz
}
\keyword{datasets}
