% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{generate}
\alias{generate}
\title{Generar data frame con 500 primeras filas, convierte direcciones
ip a dato numerico y las adiciona como nuevas columnas.}
\usage{
generate(df.osint = df)
}
\arguments{
\item{verbose}{valor TRUE}

\item{scope}{valor 500 - número de filas objetivo}
}
\value{
objeto "gn"
}
\description{
Package para generación de dataframe
}
\author{
Cristiano Dias / Luiggi Alexis Rodriguez Ruiz
}
