% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{geolocate}
\alias{geolocate}
\title{Geolocate ip address}
\usage{
geolocate(ip, format = ifelse(length(ip) == 1, "list", "dataframe"))
}
\arguments{
\item{format}{}
}
\value{
objeto "ret"
}
\description{
Función de geolocalización a partir de una dirección IP
}
