% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Grafica}
\alias{Grafica}
\title{------------------------------------OBTENER UBICACIONES EN MAPA MUNDIAL
Función que permite obtener ubicaciones en mapa mundial de Ips Atacantes/Víctimas midiendo
la variable precisión que en este caso ha sido llamada ATTACKS.}
\usage{
Grafica(geo)
}
\arguments{
\item{geo}{En este parámetro se debe ingresar el resultado de la función addIPgeolocation
A partir de ella se puede trabajar con IPs de origen o destino}
}
\value{
Gráfico de ubicaciones en mapa mundial de Ips Atacantes/Víticmas según caso origen/destino
}
\description{
------------------------------------OBTENER UBICACIONES EN MAPA MUNDIAL
Función que permite obtener ubicaciones en mapa mundial de Ips Atacantes/Víctimas midiendo
la variable precisión que en este caso ha sido llamada ATTACKS.
}
\examples{
  \dontrun{
Grafica(geo.ips)
Grafica(geo.ipd)

}

}
