% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{iconos}
\alias{iconos}
\title{------------------------------------OBTENER ICONOS PARA GRAFICO
Función que permite obtener iconos que serán empleados por la función Gráfica}
\usage{
iconos(geo)
}
\arguments{
\item{geo}{En este parámetro se debe ingresar el resultado de la función addIPgeolocation
A partir de ella se puede trabajar con IPs de origen o destino}
}
\value{
Iconos a ser empleados en gráfico.
}
\description{
------------------------------------OBTENER ICONOS PARA GRAFICO
Función que permite obtener iconos que serán empleados por la función Gráfica
}
\examples{
 \dontrun{
icons <- iconos(geo.ips)
icons <- iconos(geo.ipd)

}

}
