% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{top.paises}
\alias{top.paises}
\title{---------------------------------------PLOTEO DE PAISES QUE ORIGINAN/RECIBEN MAS ATAQUES
Función que plotea los países que generan/reciben mayor número de ataques en el mundo
de acuerdo a una muestra de 500 IPs}
\usage{
top.paises(geo)
}
\arguments{
\item{geo}{En este parámetro se debe ingresar el resultado de la función addIPgeolocation
A partir de ella se puede trabajar con IPs de origen o destino}
}
\value{
Ploteo de paísed que generan/reciben más ataques
}
\description{
---------------------------------------PLOTEO DE PAISES QUE ORIGINAN/RECIBEN MAS ATAQUES
Función que plotea los países que generan/reciben mayor número de ataques en el mundo
de acuerdo a una muestra de 500 IPs
}
\examples{
  \dontrun{
top.paises(geo.ips)
top.paises(geo.ipd)

}

}
