% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{downloadScanIO}
\alias{downloadScanIO}
\title{--------------------------------------DOWNLOAD 1
Función para descargar, descomprimir y almacenar en directorio específico el dataset
con las IPs de origen y destino de ataques
La función incluye la validación de la creación de la carpeta destino}
\usage{
downloadScanIO(n.folder = "datasets")
}
\arguments{
\item{n.folder}{Nombre del folder donde se almacenará dataset}

\item{data.url}{Introducir la url con el dataset en formato csv que queremos descargar.}

\item{filename}{Nombre de fichero que tendrá el dataset}
}
\value{
Devuelve el dataframe descargado con los datos en crudo
}
\description{
--------------------------------------DOWNLOAD 1
Función para descargar, descomprimir y almacenar en directorio específico el dataset
con las IPs de origen y destino de ataques
La función incluye la validación de la creación de la carpeta destino
}
\examples{
\dontrun{
df.tcp <- downloadScanIO()
}

}
