% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{generate}
\alias{generate}
\title{--------------------------------------SELECCIÓN MUESTRA - SCOPE
Generar data frame con 500 primeras filas de dataset TCP scans,
convierte direcciones ip a dato numérico y las adiciona como nuevas columnas.}
\usage{
generate(df.osint = df.tcp)
}
\arguments{
\item{df.osint}{variable de inicialización}

\item{verbose}{valor TRUE}

\item{scope}{valor 500 - número de filas objetivo}
}
\value{
objeto "df.muestra"
}
\description{
Package para generación de dataframe
}
\examples{
\dontrun{
df.muestra <- generate()
}

}
\author{
Cristiano Dias / Luiggi Alexis Rodriguez Ruiz
}
