% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{getColor}
\alias{getColor}
\title{------------------------------------GRAFICO DE UBICACIÓN DE FUENTES / VICTIMAS DE ATAQUES
Función de ubicación de fuentes/víctimas de ataques}
\usage{
getColor(geoips)
}
\arguments{
\item{geoips}{}
}
\description{
------------------------------------GRAFICO DE UBICACIÓN DE FUENTES / VICTIMAS DE ATAQUES
Función de ubicación de fuentes/víctimas de ataques
}
\examples{

}
